[parser] Fix #5

+++ peer/Packet15Layer.go
@@ -57,7 +57,7 @@ func (thisBitstream *extendedReader) DecodePacket15Layer(reader PacketReader, la
 	layer := NewPacket15Layer()
 
 	var err error
-	reason, err := thisBitstream.readUint16BE()
+	reason, err := thisBitstream.readUint32BE()
 	layer.Reason = int32(reason)
 	return layer, err
 }
